requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
django-allauth==65.11.1
django-appconf==1.1.0
django-ckeditor-5==0.2.18
django-crontab==0.7.1
django-imagekit==5.0.0
django-widget-tweaks==1.5.0
pilkit==3.0
pillow==11.3.0
pycparser==2.22
PyMySQL==1.1.2
sqlparse==0.5.3
tzdata==2025.2
